move validation into function.rs
